# internal/app/adapter/server/http/api/docs/swagger.yaml
basePath: /api
consumes:
- application/json
- application/x-www-form-urlencoded
definitions:
  example.ClientError:
    properties:
      errMsg:
        example: 客户端请求错误|参数校验错误
        type: string
      errNo:
        description: errNo 类型应为 int，string 仅为了表达多个错误码
        example: 10002|10003
        type: string
    type: object
  example.PermissionDenied:
    properties:
      errMsg:
        example: 暂无权限
        type: string
      errNo:
        example: 10005
        type: integer
    type: object
  example.ResourceNotFound:
    properties:
      errMsg:
        example: 资源不存在
        type: string
      errNo:
        example: 10006
        type: integer
    type: object
  example.ServerError:
    properties:
      errMsg:
        example: 服务器出错
        type: string
      errNo:
        example: 10001
        type: integer
    type: object
  example.Success:
    properties:
      data: {}
    type: object
  example.TooManyRequest:
    properties:
      errMsg:
        example: 请求过于频繁
        type: string
      errNo:
        example: 10007
        type: integer
    type: object
  example.Unauthorized:
    properties:
      errMsg:
        example: 未经授权
        type: string
      errNo:
        example: 10004
        type: integer
    type: object
...
  v1.SystemPermissionCreateRequest:
    properties:
      desc:
        type: string
      key:
        type: string
      name:
        type: string
      parentID:
        type: integer
    type: object
  v1.SystemPermissionDetailResponse:
    properties:
      desc:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      parentID:
        type: integer
    type: object
  v1.SystemPermissionInfo:
    properties:
      desc:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      parentID:
        type: integer
    type: object
  v1.SystemPermissionUpdateRequest:
    properties:
      desc:
        type: string
      id:
        type: integer
      key:
        type: string
      name:
        type: string
      parentID:
        type: integer
    type: object
...
host: localhost
info:
  contact: {}
  description: API 接口文档
  title: API 接口文档
  version: 0.0.0
paths:
...
  /v1/system-permission:
    post:
      consumes:
      - application/json
      description: 权限创建
      parameters:
      - description: 权限信息
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemPermissionCreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 权限创建
      tags:
      - 权限
    put:
      consumes:
      - application/json
      description: 权限更新
      parameters:
      - description: 权限信息
        format: string
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/v1.SystemPermissionUpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 权限更新
      tags:
      - 权限
  /v1/system-permission/{id}:
    delete:
      consumes:
      - text/plain
      description: 权限删除
      parameters:
      - description: 权限 id
        format: uint
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            $ref: '#/definitions/example.Success'
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 权限删除
      tags:
      - 权限
    get:
      consumes:
      - text/plain
      description: 权限详情
      parameters:
      - description: 权限 id
        format: uint
        in: path
        minimum: 1
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  $ref: '#/definitions/v1.SystemPermissionDetailResponse'
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 权限详情
      tags:
      - 权限
  /v1/system-permissions:
    get:
      consumes:
      - application/x-www-form-urlencoded
      description: 权限列表
      parameters:
      - description: 查询字符串
        format: string
        in: query
        name: keyword
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: 成功响应
          schema:
            allOf:
            - $ref: '#/definitions/example.Success'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/v1.SystemPermissionInfo'
                  type: array
              type: object
        "400":
          description: 客户端请求错误（code 类型应为 int，string 仅为了表达多个错误码）
          schema:
            $ref: '#/definitions/example.ClientError'
        "401":
          description: 登陆失效
          schema:
            $ref: '#/definitions/example.Unauthorized'
        "403":
          description: 没有权限
          schema:
            $ref: '#/definitions/example.PermissionDenied'
        "404":
          description: 资源不存在
          schema:
            $ref: '#/definitions/example.ResourceNotFound'
        "429":
          description: 请求过于频繁
          schema:
            $ref: '#/definitions/example.TooManyRequest'
        "500":
          description: 服务器出错
          schema:
            $ref: '#/definitions/example.ServerError'
      security:
      - Authorization: []
      summary: 权限列表
      tags:
      - 权限
...
produces:
- application/json
schemes:
- http
- https
securityDefinitions:
  Authorization:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"